circulo

+++ mainfiguras.py
@@ -32,11 +32,11 @@ while True:
         break
 
     elif opcion ==9:
-        pi = int(input("Ingresa pi: "))
-        radio = int(input("Ingresa radio: "))
+        pi = float(input("Ingresa pi: "))
+        radio = float(input("Ingresa radio: "))
         uncirculo = Circulo(pi,radio)
         print(f'El area es  {uncirculo.area()}')
-
+  
     elif opcion == 10:
         base = int(input("\nIngresa la base: "))
         altura = int(input("Ingrese la altura: "))
